refactor: [UF-246] API 모듈에서 타입 인터페이스 정의 분리 #249
