+++ src/api/services/admin/bannedWords.ts
@@ -1,73 +1,12 @@
 import { apiRequest } from '@/api/client/axios';
-
-export interface BannedWord {
-  id: number;
-  word: string;
-}
-
-export interface BannedWordsResponse {
-  statusCode: number;
-  message: string;
-  content: {
-    content: BannedWord[];
-    totalElements: number;
-    totalPages: number;
-    first: boolean;
-    last: boolean;
-    size: number;
-    number: number;
-    numberOfElements: number;
-    empty: boolean;
-    pageable: {
-      pageNumber: number;
-      pageSize: number;
-      offset: number;
-      paged: boolean;
-      unpaged: boolean;
-      sort: {
-        empty: boolean;
-        unsorted: boolean;
-        sorted: boolean;
-      };
-    };
-    sort: {
-      empty: boolean;
-      unsorted: boolean;
-      sorted: boolean;
-    };
-  };
-}
-
-export interface CreateBannedWordRequest {
-  banWord: string;
-}
-
-export interface CreateBannedWordResponse {
-  statusCode: number;
-  message: string;
-  content: BannedWord;
-}
-
-export interface DeleteBannedWordsRequest {
-  ids: number[];
-}
-
-export interface DeleteBannedWordsResponse {
-  statusCode: number;
-  message: string;
-  content: {
-    deletedIds: number[];
-    deletedCount: number;
-  };
-}
-
-export interface DeleteSingleBannedWordResponse {
-  statusCode: number;
-  message: string;
-  content: {
-    id: number;
-  };
-}
+import {
+  BannedWordsResponse,
+  CreateBannedWordRequest,
+  CreateBannedWordResponse,
+  DeleteBannedWordsRequest,
+  DeleteBannedWordsResponse,
+  DeleteSingleBannedWordResponse,
+} from '@/api/types/bannedWords';
 
 export const bannedWordsAPI = {
   // 금칙어 전체 조회 (페이지네이션)
